Backend/book_inventory/book_inventory/settings/local.py
#settings.py file to store local settings
from .base import *

#mysql database settings using python-decouple module
import os
from decouple import config

DEBUG = True
ALLOWED_HOSTS = ['*']

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.mysql',
        'NAME': config('DB_NAME'),
        'USER': config('DB_USER'),
        'PASSWORD': config('DB_PASSWORD'),
        'HOST': config('DB_HOST'),
        'PORT': config('DB_PORT'),
    }
}

# SECRET_KEY = config('SECRET_KEY')
